pyproject.toml
[project]
name = "PathRAG"
version = "0.1.0"
description = "A RAG library for path-based graph queries."
dependencies = [
    "aiohttp",
    "litellm",
    "python-dotenv",
    "numpy",
    "graspologic",
    "hnswlib",
    "nano-vectordb",
    "neo4j",
    "networkx",
    "oracledb",
    "psycopg[binary,pool]",
    "pymilvus",
    "pymongo",
    "pymysql",
    "pyvis",
    "sqlalchemy",
    "tenacity",
    "tiktoken",
    "xxhash",
]
